refactor: move purl helpers into dedicated purl.go file
